requirements.txt
biopython
numpy
matplotlib
seaborn
glob2
wget